Clarify terms in the documentation (#46)
There are many symbols not explained. It needs to be explained whether
an intercept is implicitly allowed, and whether that should be
included in the design matrix. It also seems that the `fit` API was
described in a block, but sections that follow correspond to particular
`fit` methods, so I decided to split different `fit` methods across
differerent sections of the documentation. Finally, there were several
typos and formatting errors.